mockStatic with Mockito works only with same thread so move those test to another test class
