+++ blueocean-bitbucket-pipeline/pom.xml
@@ -86,7 +86,11 @@
       <scope>test</scope>
     </dependency>
     <!-- end HttpRequest -->
-
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-inline</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 
 </project>
