+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-1.0-application-properties-q4jwO.json
@@ -1,0 +1,6 @@
+{
+  "version": "5.0.0",
+  "buildNumber": "5002000",
+  "buildDate": "1499680121775",
+  "displayName": "Bitbucket"
+}
