+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-TESTP-repos-iIrGQ.json
@@ -1,0 +1,90 @@
+{
+  "size": 2,
+  "limit": 100,
+  "isLastPage": true,
+  "values": [
+    {
+      "slug": "empty-repo-test",
+      "id": 4,
+      "name": "empty-repo-test",
+      "scmId": "git",
+      "state": "AVAILABLE",
+      "statusMessage": "Available",
+      "forkable": true,
+      "project": {
+        "key": "TESTP",
+        "id": 22,
+        "name": "testproject1",
+        "public": false,
+        "type": "NORMAL",
+        "links": {
+          "self": [
+            {
+              "href": "http://localhost:7990/projects/TESTP"
+            }
+          ]
+        }
+      },
+      "public": false,
+      "links": {
+        "clone": [
+          {
+            "href": "http://vivek@localhost:7990/scm/testp/empty-repo-test.git",
+            "name": "http"
+          },
+          {
+            "href": "ssh://git@localhost:7999/testp/empty-repo-test.git",
+            "name": "ssh"
+          }
+        ],
+        "self": [
+          {
+            "href": "http://localhost:7990/projects/TESTP/repos/empty-repo-test/browse"
+          }
+        ]
+      }
+    },
+    {
+      "slug": "pipeline-demo-test",
+      "id": 1,
+      "name": "pipeline-demo-test",
+      "scmId": "git",
+      "state": "AVAILABLE",
+      "statusMessage": "Available",
+      "forkable": true,
+      "project": {
+        "key": "TESTP",
+        "id": 22,
+        "name": "testproject1",
+        "public": false,
+        "type": "NORMAL",
+        "links": {
+          "self": [
+            {
+              "href": "http://localhost:7990/projects/TESTP"
+            }
+          ]
+        }
+      },
+      "public": false,
+      "links": {
+        "clone": [
+          {
+            "href": "http://vivek@localhost:7990/scm/testp/pipeline-demo-test.git",
+            "name": "http"
+          },
+          {
+            "href": "ssh://git@localhost:7999/testp/pipeline-demo-test.git",
+            "name": "ssh"
+          }
+        ],
+        "self": [
+          {
+            "href": "http://localhost:7990/projects/TESTP/repos/pipeline-demo-test/browse"
+          }
+        ]
+      }
+    }
+  ],
+  "start": 0
+}
