+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-branches-default-C98AD.json
@@ -1,0 +1,1 @@
+{"errors":[{"context":null,"message":"refs/heads/master is set as the default branch, but this branch does not exist","exceptionName":"com.atlassian.bitbucket.repository.NoDefaultBranchException"}]}
