+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-branches-default-EfHdI.json
@@ -1,0 +1,8 @@
+{
+  "id": "refs/heads/master",
+  "displayId": "master",
+  "type": "BRANCH",
+  "latestCommit": "0bae0ddbed2e897d3b44abc3aca9ba26e2f61710",
+  "latestChangeset": "0bae0ddbed2e897d3b44abc3aca9ba26e2f61710",
+  "isDefault": true
+}
