+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-browse-Jenkinsfile-YItmp.json
@@ -1,0 +1,1 @@
+{"lines":[{"text":"node{"},{"text":"  echo 'hello world'"},{"text":"}"}],"start":0,"size":3,"isLastPage":true}
