+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-browse-Jenkinsfile-put-update.json
@@ -1,0 +1,45 @@
+{
+  "committer" : {
+    "name" : "vivek",
+    "id" : 1,
+    "displayName" : "Vivek Pandey",
+    "links" : {
+      "self" : [
+        {
+          "href" : "http://localhost:7990/users/vivek"
+        }
+      ]
+    },
+    "emailAddress" : "vivek@example.com",
+    "slug" : "vivek",
+    "type" : "NORMAL",
+    "active" : true
+  },
+  "message" : "Updated using file-edit REST API",
+  "author" : {
+    "type" : "NORMAL",
+    "slug" : "vivek",
+    "emailAddress" : "vivek@example.com",
+    "active" : true,
+    "id" : 1,
+    "name" : "vivek",
+    "links" : {
+      "self" : [
+        {
+          "href" : "http://localhost:7990/users/vivek"
+        }
+      ]
+    },
+    "displayName" : "Vivek Pandey"
+  },
+  "id" : "e587b620844b1b230783976f00cfb8383488aeca",
+  "displayId" : "e587b620844",
+  "authorTimestamp" : 1497281170000,
+  "committerTimestamp" : 1497281170000,
+  "parents" : [
+    {
+      "displayId" : "a77840d4108",
+      "id" : "a77840d4108db2befe6c616723eb3f4485af5d24"
+    }
+  ]
+}
