+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-browse-README-put-update.json
@@ -1,0 +1,45 @@
+{
+  "id" : "a77840d4108db2befe6c616723eb3f4485af5d24",
+  "authorTimestamp" : 1497281017000,
+  "committerTimestamp" : 1497281017000,
+  "author" : {
+    "slug" : "vivek",
+    "name" : "vivek",
+    "links" : {
+      "self" : [
+        {
+          "href" : "http://localhost:7990/users/vivek"
+        }
+      ]
+    },
+    "id" : 1,
+    "active" : true,
+    "displayName" : "Vivek Pandey",
+    "type" : "NORMAL",
+    "emailAddress" : "vivek@example.com"
+  },
+  "committer" : {
+    "emailAddress" : "vivek@example.com",
+    "type" : "NORMAL",
+    "id" : 1,
+    "active" : true,
+    "displayName" : "Vivek Pandey",
+    "name" : "vivek",
+    "slug" : "vivek",
+    "links" : {
+      "self" : [
+        {
+          "href" : "http://localhost:7990/users/vivek"
+        }
+      ]
+    }
+  },
+  "displayId" : "a77840d4108",
+  "parents" : [
+    {
+      "displayId" : "c246b778fb2",
+      "id" : "c246b778fb2496260f7942d4f8d0ce364eba4ff8"
+    }
+  ],
+  "message" : "another commit"
+}
