+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-browse-README.md-MXlpn.md
@@ -1,0 +1,1 @@
+{"lines":[{"text":"This is test content in new file"}],"start":0,"size":1,"isLastPage":true}
