+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-empty-repo-not-found-head.json
@@ -1,0 +1,9 @@
+{
+  "errors" : [
+    {
+      "exceptionName" : "com.atlassian.bitbucket.repository.NoDefaultBranchException",
+      "context" : null,
+      "message" : "refs/heads/master is set as the default branch, but this branch does not exist"
+    }
+  ]
+}
