+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-projects-TESTP-FuH7O.json
@@ -1,0 +1,1 @@
+{"key":"TESTP","id":22,"name":"testproject1","public":false,"type":"NORMAL","links":{"self":[{"href":"http://localhost:7990/projects/TESTP"}]}}
