+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/__files/body-repos-pipeline-demo-test-PcCw8.json
@@ -1,0 +1,1 @@
+{"slug":"pipeline-demo-test","id":1,"name":"pipeline-demo-test","scmId":"git","state":"AVAILABLE","statusMessage":"Available","forkable":true,"project":{"key":"TESTP","id":22,"name":"testproject1","public":false,"type":"NORMAL","links":{"self":[{"href":"http://localhost:7990/projects/TESTP"}]}},"public":false,"links":{"clone":[{"href":"http://vivek@localhost:7990/scm/testp/pipeline-demo-test.git","name":"http"},{"href":"ssh://git@localhost:7999/testp/pipeline-demo-test.git","name":"ssh"}],"self":[{"href":"http://localhost:7990/projects/TESTP/repos/pipeline-demo-test/browse"}]}}
