+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-TESTP-repos-iIrGQ.json
@@ -1,0 +1,24 @@
+{
+  "id" : "4abc8179-5c7e-3bcb-afef-804c517468be",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/?start=0&limit=100",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-TESTP-repos-iIrGQ.json",
+    "headers" : {
+      "X-AREQUESTID" : "@1M44HP4x855x1960x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Mon, 12 Jun 2017 14:15:00 GMT"
+    }
+  },
+  "uuid" : "4abc8179-5c7e-3bcb-afef-804c517468be"
+}
