+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-branches-default-5.6.0.json
@@ -1,0 +1,24 @@
+{
+  "id" : "3cb8ad8e-bf64-3779-b249-1c1ca33c783a",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/empty1/branches/default",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 204,
+    "bodyFileName" : "body-branches-default-C98AD.json",
+    "headers" : {
+      "X-AREQUESTID" : "@190KU8Bx1085x620x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Wed, 21 Jun 2017 18:05:02 GMT"
+    }
+  },
+  "uuid" : "3cb8ad8e-bf64-3779-b249-1c1ca33c783a"
+}
