+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-browse-Jenkinsfile-z65wE.json
@@ -1,0 +1,24 @@
+{
+  "id" : "018b0312-abb1-3c5e-a76c-72aade97da4f",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/pipeline-demo-test/browse/Jenkinsfile?at=c246b778fb2496260f7942d4f8d0ce364eba4ff8&start=0&limit=500",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-browse-Jenkinsfile-z65wE.json",
+    "headers" : {
+      "X-AREQUESTID" : "@190KU8Bx712x424x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "private, no-transform, max-age=2419200",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Tue, 20 Jun 2017 11:52:55 GMT"
+    }
+  },
+  "uuid" : "018b0312-abb1-3c5e-a76c-72aade97da4f"
+}
