+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-browse-README-head-master.json
@@ -1,0 +1,23 @@
+{
+  "id" : "fcd61006-ea0c-3033-ac53-50210cdbb644",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/pipeline-demo-test/browse/README.md?at=refs%2Fheads%2Fmaster",
+    "method" : "HEAD"
+  },
+  "response" : {
+    "status" : 404,
+    "headers" : {
+      "X-AREQUESTID" : "@1M44HP4x855x1961x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Mon, 12 Jun 2017 14:15:28 GMT"
+    }
+  },
+  "uuid" : "fcd61006-ea0c-3033-ac53-50210cdbb644"
+}
