+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-browse-README.md-y3xys.json
@@ -1,0 +1,24 @@
+{
+  "id" : "3f0e3cb8-c58b-3a1a-92c8-6bb1bb02765f",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/pipeline-demo-test/browse/README.md?at=a77840d4108db2befe6c616723eb3f4485af5d24&start=0&limit=500",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-browse-README.md-y3xys.json",
+    "headers" : {
+      "X-AREQUESTID" : "@190KU8Bx712x425x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "private, no-transform, max-age=2419200",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Tue, 20 Jun 2017 11:52:55 GMT"
+    }
+  },
+  "uuid" : "3f0e3cb8-c58b-3a1a-92c8-6bb1bb02765f"
+}
