+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-empty-repo-put-newfile-README.json
@@ -1,0 +1,22 @@
+{
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/empty-repo-test/browse/README.md",
+    "method" : "PUT"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName": "body-browse-README-empty-repo-test-update.json",
+    "headers" : {
+      "X-AREQUESTID": "@1M44HP4x873x2017x0",
+      "X-ASEN": "SEN-L9817337",
+      "X-AUSERID": "1",
+      "X-AUSERNAME": "vivek",
+      "Cache-Control": "no-cache, no-transform",
+      "Vary": "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding": "chunked",
+      "Content-Type": "application/json;charset=UTF-8",
+      "X-Content-Type-Options": "nosniff",
+      "Date": "Mon, 12 Jun 2017 14:33:18 GMT"
+    }
+  }
+}
