+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-pipeline-demo-test-branches-Ccvaw.json
@@ -1,0 +1,24 @@
+{
+  "id" : "21f37681-abff-31fc-800a-38c1eb44d6a1",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/pipeline-demo-test/branches/?filterText=master",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-pipeline-demo-test-branches-Ccvaw.json",
+    "headers" : {
+      "X-AREQUESTID" : "@1M44HP4x856x1962x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Mon, 12 Jun 2017 14:16:57 GMT"
+    }
+  },
+  "uuid" : "21f37681-abff-31fc-800a-38c1eb44d6a1"
+}
