+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-projects-TESTP-FuH7O.json
@@ -1,0 +1,24 @@
+{
+  "id" : "767c559f-1935-31fb-be42-6a81bb26ccaf",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-projects-TESTP-FuH7O.json",
+    "headers" : {
+      "X-AREQUESTID" : "@190KU8Bx703x607x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Wed, 21 Jun 2017 11:43:22 GMT"
+    }
+  },
+  "uuid" : "767c559f-1935-31fb-be42-6a81bb26ccaf"
+}
