+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-redirect-test.json
@@ -1,0 +1,23 @@
+{
+  "id" : "d5434dae-88c5-3cdd-a506-11adb708a1ad",
+  "request" : {
+    "url" : "/rest/api/1.0/test-redirect",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 302,
+    "bodyFileName" : "body-1.0-application-properties-q4jwO.json",
+    "headers" : {
+      "X-AREQUESTID" : "@1962TV2x594x11x0",
+      "X-ASEN" : "SEN-L9817337",
+      "Location" : "http://localhost:7990/bitbucket/rest/api/1.0/redirect-test-success",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Wed, 12 Jul 2017 09:54:11 GMT"
+    }
+  },
+  "uuid" : "d5434dae-88c5-3cdd-a506-11adb708a1ad"
+}
