+++ blueocean-bitbucket-pipeline/src/test/resources/api/server_wrong_version/mappings/mapping-repos-pipeline-demo-test-PcCw8.json
@@ -1,0 +1,24 @@
+{
+  "id" : "95be5f1c-6099-3d8f-a28a-b63da5386da6",
+  "request" : {
+    "url" : "/rest/api/1.0/projects/TESTP/repos/pipeline-demo-test",
+    "method" : "GET"
+  },
+  "response" : {
+    "status" : 200,
+    "bodyFileName" : "body-repos-pipeline-demo-test-PcCw8.json",
+    "headers" : {
+      "X-AREQUESTID" : "@190KU8Bx703x608x0",
+      "X-ASEN" : "SEN-L9817337",
+      "X-AUSERID" : "1",
+      "X-AUSERNAME" : "vivek",
+      "Cache-Control" : "no-cache, no-transform",
+      "Vary" : "X-AUSERNAME,Accept-Encoding",
+      "Transfer-Encoding" : "chunked",
+      "X-Content-Type-Options" : "nosniff",
+      "Content-Type" : "application/json;charset=UTF-8",
+      "Date" : "Wed, 21 Jun 2017 11:43:22 GMT"
+    }
+  },
+  "uuid" : "95be5f1c-6099-3d8f-a28a-b63da5386da6"
+}
